+++ src/main/java/org/example/modelo/Libro.java
@@ -8,7 +8,7 @@ public class Libro {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(nullable = false)
     private String titulo;
@@ -39,47 +39,58 @@ public class Libro {
     }
 
     // Getters y Setters
-    public Long getId() {
+    public int getId() {
+
         return id;
     }
 
     public String getTitulo() {
+
         return titulo;
     }
 
     public void setTitulo(String titulo) {
+
         this.titulo = titulo;
     }
 
     public String getIsbn() {
+
         return isbn;
     }
 
     public void setIsbn(String isbn) {
+
         this.isbn = isbn;
     }
 
     public String getEditorial() {
+
         return editorial;
     }
 
     public void setEditorial(String editorial) {
+
         this.editorial = editorial;
     }
 
     public int getAnioPublicacion() {
+
         return anioPublicacion;
     }
 
     public void setAnioPublicacion(int anioPublicacion) {
+
         this.anioPublicacion = anioPublicacion;
     }
 
     public Autor getAutor() {
+
         return autor;
     }
 
     public void setAutor(Autor autor) {
+
         this.autor = autor;
     }
