+++ src/main/java/org/example/modelo/Socio.java
@@ -8,7 +8,7 @@ public class Socio {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(nullable = false)
     private String nombre;
@@ -30,31 +30,38 @@ public class Socio {
     }
 
     // Getters y Setters
-    public Long getId() {
+    public int getId() {
+
         return id;
     }
 
     public String getNombre() {
+
         return nombre;
     }
 
     public void setNombre(String nombre) {
+
         this.nombre = nombre;
     }
 
     public String getDireccion() {
+
         return direccion;
     }
 
     public void setDireccion(String direccion) {
+
         this.direccion = direccion;
     }
 
     public String getTelefono() {
+
         return telefono;
     }
 
     public void setTelefono(String telefono) {
+
         this.telefono = telefono;
     }
